Get AIC

get.AIC retrieves Akaike information criteria from a fitted hydroState model object or all models.

Usage

get.AIC(model)

Arguments

model

fitted hydroState model object.

Value

AIC value of a single model or a list variable of AIC values for al models

Details

get.AIC

The AIC is the negative log-likelihood of the model plus a penalty for model parameters. This function can be performed on a single model or a selection of models to find the lowest AIC of the set.

Examples

# Load fitted model
data(model.annual.fitted.221201)

## AIC of a single model
get.AIC(model.annual.fitted.221201)
#> [1] 89.81635

## Lowest AIC of a model set
get.AIC(all.models.annual.fitted.407211)
#> ... The following model has a state with only 1  time points:  model.3State.normal.boxcox.AR2.a0std
#> ... The following model has a state with only 1  time points:  model.3State.normal.boxcox.AR3.a0std
#>                                                 AIC nParameters  AIC.weights
#> model.2State.gamma.boxcox.AR0.a0std        79.36469           8 3.890331e-01
#> model.2State.gamma.boxcox.AR2.a0std        79.39221          10 3.837172e-01
#> model.2State.gamma.boxcox.AR3.a0std        81.39029          11 1.412971e-01
#> model.2State.gamma.boxcox.AR1.a0std        82.48915           9 8.156764e-02
#> model.1State.gamma.boxcox.AR2.a0std        89.13483           5 2.940536e-03
#> model.1State.gamma.boxcox.AR3.a0std        91.83986           6 7.603942e-04
#> model.1State.gamma.boxcox.AR1.a0std        92.18365           4 6.403015e-04
#> model.3State.gamma.boxcox.AR0.a0std        98.72607          15 2.430621e-05
#> model.2State.truc.normal.boxcox.AR0.a0std 100.52439           8 9.890434e-06
#> model.3State.gamma.boxcox.AR1.a0std       103.30837          16 2.458567e-06
#> model.2State.truc.normal.boxcox.AR1.a0std 103.70400           9 2.017312e-06
#> model.1State.gamma.boxcox.AR0.a0std       104.00988           3 1.731217e-06
#> model.1State.truc.normal.boxcox.AR2.a0std 105.46604           5 8.358910e-07
#> model.2State.truc.normal.boxcox.AR2.a0std 105.63994          10 7.662811e-07
#> model.2State.normal.boxcox.AR0.a0std      106.58864           8 4.768500e-07
#> model.3State.gamma.boxcox.AR2.a0std       107.10237          17 3.688297e-07
#> model.1State.truc.normal.boxcox.AR3.a0std 107.52736           6 2.982222e-07
#> model.2State.truc.normal.boxcox.AR3.a0std 109.00463          11 1.424801e-07
#> model.3State.gamma.boxcox.AR3.a0std       109.42471          18 1.154880e-07
#> model.1State.truc.normal.boxcox.AR1.a0std 109.71836           4 9.971724e-08
#> model.2State.normal.boxcox.AR1.a0std      109.75382           9 9.796489e-08
#> model.1State.normal.boxcox.AR2.a0std      111.32092           5 4.474858e-08
#> model.1State.truc.normal.boxcox.AR0.a0std 112.84279           3 2.090786e-08
#> model.2State.normal.boxcox.AR2.a0std      112.94247          10 1.989136e-08
#> model.1State.normal.boxcox.AR3.a0std      113.48421           6 1.517142e-08
#> model.1State.normal.boxcox.AR1.a0std      113.93972           4 1.208134e-08
#> model.2State.normal.boxcox.AR3.a0std      114.30162          11 1.008158e-08
#> model.1State.normal.boxcox.AR0.a0std      116.25319           3 3.799707e-09
#> model.3State.truc.normal.boxcox.AR0.a0std 116.54611          15 3.282036e-09
#> model.3State.truc.normal.boxcox.AR1.a0std 120.87487          16 3.768473e-10
#> model.3State.truc.normal.boxcox.AR2.a0std 123.18625          17 1.186467e-10
#> model.3State.truc.normal.boxcox.AR3.a0std 128.25831          18 9.394473e-12
#> model.3State.normal.boxcox.AR0.a0std      130.17241          15 3.607707e-12
#> model.3State.normal.boxcox.AR1.a0std      134.85168          16 3.476474e-13