Get pseudo residuals

The pseudo residuals were derived from the conditional probabilities of the observations. At each time-step, the pseudo residual is the probability of an observation occurring given the prior observations and latter observations.

Usage

get.residuals(model)

Arguments

model

fitted hydroState model object.

Value

Data frame of residuals for each time-step

Details

get.residuals

get.residuals retrieves residuals from the fitted model and exports them as a data frame.

Examples

# Load fitted model
data(model.annual.fitted.221201)

## Get residuals in a dataframe
get.residuals(model = model.annual.fitted.221201)
#>    year      flow precipitation Qhat.flow Qhat.precipitation unif.pseuou.resid
#> 1  1923 125.09053      961.6729  4.837000           961.6729       0.282360012
#> 2  1924  63.94195      795.7554  4.173494           795.7554       0.205443625
#> 3  1925 184.41702     1045.1675  5.222607          1045.1675       0.388304359
#> 4  1926  71.62613      749.1795  4.285325           749.1795       0.530903701
#> 5  1927 130.10813      981.1049  4.876022           981.1049       0.256435045
#> 6  1928 269.01408      969.8059  5.598474           969.8059       0.964239123
#> 7  1929        NA      928.4907        NA           928.4907                NA
#> 8  1930        NA      932.5074        NA           932.5074                NA
#> 9  1931        NA      865.8296        NA           865.8296                NA
#> 10 1932        NA      972.0635        NA           972.0635                NA
#> 11 1933        NA     1253.4182        NA          1253.4182                NA
#> 12 1934        NA     1457.7865        NA          1457.7865                NA
#> 13 1935        NA      947.9338        NA           947.9338                NA
#> 14 1936        NA      869.2778        NA           869.2778                NA
#> 15 1937        NA      936.5729        NA           936.5729                NA
#> 16 1938        NA      872.1881        NA           872.1881                NA
#> 17 1939        NA      870.5235        NA           870.5235                NA
#> 18 1940        NA      806.5427        NA           806.5427                NA
#> 19 1941        NA      672.8050        NA           672.8050                NA
#> 20 1942        NA      926.1249        NA           926.1249                NA
#> 21 1943        NA      695.4942        NA           695.4942                NA
#> 22 1944        NA      824.9850        NA           824.9850                NA
#> 23 1945        NA      856.2976        NA           856.2976                NA
#> 24 1946        NA      953.9565        NA           953.9565                NA
#> 25 1947 106.87001      904.7650  4.680927           904.7650       0.353287786
#> 26 1948 152.83273      969.1042  5.035866           969.1042       0.497172893
#> 27 1949 269.40076     1081.6194  5.599905          1081.6194       0.697165777
#> 28 1950 448.62075     1236.2350  6.108404          1236.2350       0.651928408
#> 29 1951 388.09704     1083.6728  5.963829          1083.6728       0.952628956
#> 30 1952        NA     1497.3913        NA          1497.3913                NA
#> 31 1953        NA      922.0591        NA           922.0591                NA
#> 32 1954        NA      714.8487        NA           714.8487                NA
#> 33 1955        NA      934.8080        NA           934.8080                NA
#> 34 1956        NA     1118.4126        NA          1118.4126                NA
#> 35 1957 159.99592      985.0376  5.081379           985.0376       0.489706408
#> 36 1958  60.15846      718.3842  4.113468           718.3842       0.456390805
#> 37 1959 255.76358     1235.8035  5.548156          1235.8035       0.082635987
#> 38 1960 254.81734     1093.9330  5.544464          1093.9330       0.583992796
#> 39 1961 392.20071     1340.7397  5.974320          1340.7397       0.111047954
#> 40 1962 123.98650      802.7521  4.828206           802.7521       0.888813453
#> 41 1963 275.55938     1067.5055  5.622426          1067.5055       0.776686704
#> 42 1964 109.58464      875.7446  4.705781           875.7446       0.521384141
#> 43 1965        NA      846.4930        NA           846.4930                NA
#> 44 1966 285.17953     1219.3717  5.656619          1219.3717       0.317565507
#> 45 1967  83.88916      705.8440  4.441346           705.8440       0.923285767
#> 46 1968  35.63002      880.7425  3.600868           880.7425       0.005046797
#> 47 1969 156.71943     1074.2433  5.060818          1074.2433       0.523658725
#> 48 1970        NA     1516.1403        NA          1516.1403                NA
#> 49 1971        NA      753.4645        NA           753.4645                NA
#> 50 1972  31.71113      514.5234  3.487715           514.5234       0.689934534
#> 51 1973  68.07280      864.5321  4.235161           864.5321       0.118714928
#> 52 1974 452.44566     1348.9024  6.116875          1348.9024       0.227030319
#> 53 1975 331.58509     1053.5235  5.806896          1053.5235       0.935192909
#> 54 1976 251.36386      992.4163  5.530872           992.4163       0.906051110
#> 55 1977 126.59230      791.1651  4.848840           791.1651       0.921066231
#> 56 1978 546.16352     1311.3833  6.304748          1311.3833       0.570963360
#> 57 1979  64.08882      709.9309  4.175753           709.9309       0.594340275
#> 58 1980  21.94957      615.9869  3.133299           615.9869       0.026384968
#> 59 1981  35.73008      645.2430  3.603596           645.2430       0.240714515
#> 60 1982  24.04471      614.2559  3.220663           614.2559       0.049092189
#> 61 1983 234.84171     1057.5951  5.463161          1057.5951       0.655581349
#> 62 1984 306.71210      994.5065  5.729165           994.5065       0.973381871
#> 63 1985 651.97575     1314.2949  6.481540          1314.2949       0.758630857
#> 64 1986  98.10944      664.3635  4.596225           664.3635       0.979579972
#> 65 1987  52.39369      655.7308  3.977693           655.7308       0.558221010
#> 66 1988 135.15181      932.1170  4.913771           932.1170       0.503157793
#> 67 1989 257.87835     1123.2078  5.556358          1123.2078       0.459140897
#> 68 1990 172.44735      893.0441  5.155874           893.0441       0.891867083
#> 69 1991 188.67362     1070.6112  5.245305          1070.6112       0.307760997
#> 70 1992 189.19196      998.9060  5.248034           998.9060       0.629434226
#> 71 1993  88.38179      902.7844  4.492917           902.7844       0.160873059
#> 72 1994  87.99908      838.0450  4.488626           838.0450       0.402072977
#> 73 1995 139.66260     1028.2866  4.946364          1028.2866       0.183716646
#> 74 1996  61.28702      754.2106  4.131753           754.2106       0.370900387
#> 75 1997  51.92631      679.9007  3.968901           679.9007       0.557207831
#> 76 1998 135.79290      999.5150  4.918468           999.5150       0.487249117
#> 77 1999  42.27925      751.4960  3.767673           751.4960       0.397844091
#> 78 2000  44.45604      906.4195  3.816746           906.4195       0.209180324
#> 79 2001 151.32080     1188.6574  5.025989          1188.6574       0.401022743
#> 80 2002  99.88997      694.2971  4.614030           694.2971       0.983078681
#> 81 2003  24.55597      753.2511  3.240871           753.2511       0.256132490
#> 82 2004  21.93022      873.0992  3.132456           873.0992       0.058297582
#> 83 2005  47.45737      841.9714  3.880684           841.9714       0.451713013
#> 84 2006  72.10276      821.5185  4.291866           821.5185       0.767453030
#> 85 2007 139.71488      956.2435  4.946736           956.2435       0.854382702
#> 86 2008  18.05987      648.0498  2.947585           648.0498       0.292415514
#> 87 2009  16.74858      837.1073  2.876306           837.1073       0.036696607
#> 88 2010  21.36819      804.4517  3.107640           804.4517       0.117427502
#> 89 2011 173.88331     1188.6321  5.164119          1188.6321       0.505455717
#> 90 2012 224.61359     1219.1642  5.418824          1219.1642       0.601268000
#> 91 2013 123.83128      966.0427  4.826963           966.0427       0.788537296
#> 92 2014 132.86532     1275.3471  4.896834          1275.3471       0.051273823
#> 93 2015 245.60145     1144.0182  5.507773          1144.0182       0.687090854
#> 94 2016  97.54142      832.5498  4.590477           832.5498       0.786505168
#>    norm.pseudo.resid
#> 1       -0.575844898
#> 2       -0.822333264
#> 3       -0.283741201
#> 4        0.077541726
#> 5       -0.654375229
#> 6        1.802150344
#> 7                 NA
#> 8                 NA
#> 9                 NA
#> 10                NA
#> 11                NA
#> 12                NA
#> 13                NA
#> 14                NA
#> 15                NA
#> 16                NA
#> 17                NA
#> 18                NA
#> 19                NA
#> 20                NA
#> 21                NA
#> 22                NA
#> 23                NA
#> 24                NA
#> 25      -0.376459158
#> 26      -0.007086565
#> 27       0.516266278
#> 28       0.390532018
#> 29       1.670896675
#> 30                NA
#> 31                NA
#> 32                NA
#> 33                NA
#> 34                NA
#> 35      -0.025805072
#> 36      -0.109530655
#> 37      -1.387557031
#> 38       0.212118730
#> 39      -1.220973886
#> 40       1.220242149
#> 41       0.761051025
#> 42       0.053627787
#> 43                NA
#> 44      -0.474517370
#> 45       1.427525551
#> 46      -2.572606359
#> 47       0.059338434
#> 48                NA
#> 49                NA
#> 50       0.495664791
#> 51      -1.181435262
#> 52      -0.748662522
#> 53       1.515625076
#> 54       1.316823544
#> 55       1.412279984
#> 56       0.178827336
#> 57       0.238724200
#> 58      -1.936798855
#> 59      -0.704006011
#> 60      -1.653720197
#> 61       0.400433436
#> 62       1.932999509
#> 63       0.701905199
#> 64       2.045150168
#> 65       0.146460361
#> 66       0.007915496
#> 67      -0.102598297
#> 68       1.236518660
#> 69      -0.502206895
#> 70       0.330355254
#> 71      -0.990876028
#> 72      -0.247985075
#> 73      -0.901291611
#> 74      -0.329469591
#> 75       0.143893791
#> 76      -0.031967168
#> 77      -0.258931385
#> 78      -0.809268627
#> 79      -0.250700743
#> 80       2.121941648
#> 81      -0.655314976
#> 82      -1.569226532
#> 83      -0.121334588
#> 84       0.730484737
#> 85       1.055417121
#> 86      -0.546341772
#> 87      -1.790377690
#> 88      -1.187945188
#> 89       0.013675882
#> 90       0.256630537
#> 91       0.801356297
#> 92      -1.632626160
#> 93       0.487621038
#> 94       0.794353505